Had GPT Astra attempt to fix the scrolling bug that has been plaguing me for months. It did it!

+++ include/CPU/CPUState.h
@@ -25,7 +25,7 @@ struct CPUState {
   CPUState(uint16_t pc, const OpCode& op, const std::vector<uint8_t>& opBytes,
            const AddressResolveInfo& addrInfo, const uint8_t& valueAtAddr,
            uint8_t A,
-           uint8_t X, uint8_t Y, uint8_t P, uint8_t SP, int ppuX, int ppuY,
+           uint8_t X, uint8_t Y, uint8_t P, uint8_t SP, uint16_t ppuX, uint16_t ppuY,
            uint64_t cycles)
       : pc(pc),
         op(op),
